Move validationRules to SearchHolidays
